feat(booking) : Booking any service by customer with proper validation

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\API\AuthController;
+use App\Http\Controllers\API\BookingController;
 use App\Http\Controllers\API\ServiceController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -9,7 +10,6 @@ use Illuminate\Support\Facades\Route;
 Route::controller(AuthController::class)->group(function(){
     Route::post('register', 'register');
     Route::post('login', 'login');
-    // Route::post('logout', 'logout');
 });
 
 
@@ -18,10 +18,13 @@ Route::middleware('auth:sanctum')->group(function () {
 
     Route::prefix('services')->group(function () {
         Route::controller(ServiceController::class)->group(function () {
-            Route::get('/', 'index');
-            Route::post('/', 'store');
-            Route::put('/{id}', 'update');
-            Route::delete('/{id}', 'destroy');
+            Route::get('/', 'index'); //Customer
+            Route::post('/', 'store'); //Admin
+            Route::put('/{id}', 'update'); //Admin
+            Route::delete('/{id}', 'destroy'); //Admin
         });
     });
+
+
+    Route::post('/bookings', [BookingController::class, 'store']);
 });
